Allow onetun to be used as a library
