+++ .github/workflows/build.yml
@@ -26,6 +26,11 @@ jobs:
         with:
           command: check
 
+      - name: Run cargo check without default features
+        uses: actions-rs/cargo@v1
+        with:
+          command: check --no-default-features
+
   test:
     name: Test Suite
     runs-on: ubuntu-latest
